requirements.txt
streamlit
seaborn
jupyter
numpy
matplotlib.pyplot
os
google